feat: Create new client from Auth0 login
